--- a/central/networkgraph/flow/datastore/internal/store/postgres/store.go
+++ b/central/networkgraph/flow/datastore/internal/store/postgres/store.go
@@ -50,12 +50,12 @@ const (
 
 	deleteStmt         = "DELETE FROM network_flows_v2 WHERE Props_SrcEntity_Type = $1 AND Props_SrcEntity_Id = $2 AND Props_DstEntity_Type = $3 AND Props_DstEntity_Id = $4 AND Props_DstPort = $5 AND Props_L4Protocol = $6 AND ClusterId = $7"
 	deleteStmtWithTime = "DELETE FROM network_flows_v2 WHERE Props_SrcEntity_Type = $1 AND Props_SrcEntity_Id = $2 AND Props_DstEntity_Type = $3 AND Props_DstEntity_Id = $4 AND Props_DstPort = $5 AND Props_L4Protocol = $6 AND ClusterId = $7 AND LastSeenTimestamp = $8"
-	walkStmt           = "SELECT nf.Props_SrcEntity_Type, nf.Props_SrcEntity_Id, nf.Props_DstEntity_Type, nf.Props_DstEntity_Id, nf.Props_DstPort, nf.Props_L4Protocol, nf.LastSeenTimestamp, nf.ClusterId::text FROM network_flows_v2 nf " + joinStmt + " WHERE nf.ClusterId = $1"
+	walkStmt           = "SELECT nf.Props_SrcEntity_Type, nf.Props_SrcEntity_Id, nf.Props_DstEntity_Type, nf.Props_DstEntity_Id, nf.Props_DstPort, nf.Props_L4Protocol, nf.LastSeenTimestamp, nf.ClusterId::text FROM %s nf " + joinStmt + " WHERE nf.ClusterId = $1"
 
 	// These mimic how the RocksDB version of the flow store work
 	getSinceStmt = `SELECT nf.Props_SrcEntity_Type, nf.Props_SrcEntity_Id, nf.Props_DstEntity_Type, 
 	nf.Props_DstEntity_Id, nf.Props_DstPort, nf.Props_L4Protocol, nf.LastSeenTimestamp, nf.ClusterId::text 
-	FROM network_flows_v2 nf ` + joinStmt +
+	FROM %s nf ` + joinStmt +
 		` WHERE (nf.LastSeenTimestamp >= $1 OR nf.LastSeenTimestamp IS NULL) AND nf.ClusterId = $2`
 	deleteSrcDeploymentStmt = "DELETE FROM network_flows_v2 WHERE ClusterId = $1 AND Props_SrcEntity_Type = 1 AND Props_SrcEntity_Id = $2"
 	deleteDstDeploymentStmt = "DELETE FROM network_flows_v2 WHERE ClusterId = $1 AND Props_DstEntity_Type = 1 AND Props_DstEntity_Id = $2"
@@ -420,10 +420,10 @@ func (s *flowStoreImpl) retryableGetAllFlows(ctx context.Context, since *types.T
 
 	// handling case when since is nil.  Assumption is we want everything in that case vs when date is not null
 	if since == nil {
-		partitionWalkStmt := fmt.Sprintf(walkStmt, s.partitionName)
+		partitionWalkStmt := fmt.Sprintf(walkStmt, s.partitionName, s.partitionName)
 		rows, err = s.db.Query(ctx, partitionWalkStmt, s.clusterID)
 	} else {
-		partitionSinceStmt := fmt.Sprintf(getSinceStmt, s.partitionName)
+		partitionSinceStmt := fmt.Sprintf(getSinceStmt, s.partitionName, s.partitionName)
 		rows, err = s.db.Query(ctx, partitionSinceStmt, pgutils.NilOrTime(since), s.clusterID)
 	}
 	if err != nil {
@@ -457,10 +457,10 @@ func (s *flowStoreImpl) retryableGetMatchingFlows(ctx context.Context, pred func
 
 	// handling case when since is nil.  Assumption is we want everything in that case vs when date is not null
 	if since == nil {
-		partitionWalkStmt := fmt.Sprintf(walkStmt, s.partitionName)
+		partitionWalkStmt := fmt.Sprintf(walkStmt, s.partitionName, s.partitionName)
 		rows, err = s.db.Query(ctx, partitionWalkStmt, s.clusterID)
 	} else {
-		partitionSinceStmt := fmt.Sprintf(getSinceStmt, s.partitionName)
+		partitionSinceStmt := fmt.Sprintf(getSinceStmt, s.partitionName, s.partitionName)
 		rows, err = s.db.Query(ctx, partitionSinceStmt, pgutils.NilOrTime(since), s.clusterID)
 	}
 
@@ -575,7 +575,7 @@ func (s *flowStoreImpl) retryableRemoveMatchingFlows(ctx context.Context, keyMat
 	// in Postgres we cannot fully do this work in SQL.  Additionally, there may be issues with the synchronization
 	// of when flow is created vs a deployment deleted that may also make that problematic.
 	if keyMatchFn != nil {
-		partitionWalkStmt := fmt.Sprintf(walkStmt, s.partitionName)
+		partitionWalkStmt := fmt.Sprintf(walkStmt, s.partitionName, s.partitionName)
 		rows, err := conn.Query(ctx, partitionWalkStmt, s.clusterID)
 		if err != nil {
 			return err