--- a/migrator/migrations/m_172_to_m_173_network_flows_partition/stores/updated/store.go
+++ b/migrator/migrations/m_172_to_m_173_network_flows_partition/stores/updated/store.go
@@ -178,7 +178,7 @@ func (s *flowStoreImpl) readRows(rows pgx.Rows, pred func(*storage.NetworkFlowPr
 	}
 
 	log.Debugf("Read returned %d flows", len(flows))
-	return flows, nil
+	return flows, rows.Err()
 }
 
 // GetAllFlows returns the object, if it exists from the store, timestamp and error