--- a/pkg/postgres/paths.go
+++ b/pkg/postgres/paths.go
@@ -2,7 +2,9 @@ package postgres
 
 // CreateStmts holds the model and statements for creating sql table.
 type CreateStmts struct {
-	GormModel interface{}
-	Children  []*CreateStmts
-	PostStmts []string
+	GormModel       interface{}
+	Children        []*CreateStmts
+	PostStmts       []string
+	Partition       bool
+	PartitionCreate string
 }